Refactoring scheduling into the new paradigm
